requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coverage==7.0.5
flasgger==0.9.5
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.46
urllib3==1.26.14
Werkzeug==2.2.2
